# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News portal",
        "description": "Documentation for the NewsAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/userSettings": {
            "get": {
                "tags": [
                    "User Settings"
                ],
                "summary": "Show the form for editing user settings.",
                "description": "Returns the form for editing user settings including preferred category, country, and language.",
                "operationId": "userSettings.edit",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "allLanguages": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "allCategories": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "allCountries": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "User Settings"
                ],
                "summary": "Update the user settings.",
                "description": "Updates the user settings including preferred category, country, and language.",
                "operationId": "userSettings.update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category",
                                    "country",
                                    "language"
                                ],
                                "properties": {
                                    "category": {
                                        "description": "Preferred news category",
                                        "type": "string"
                                    },
                                    "country": {
                                        "description": "Preferred news country",
                                        "type": "string"
                                    },
                                    "language": {
                                        "description": "Preferred news language",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Successful operation. Redirects to the user settings edit page.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User settings updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}